pull the mouse click handler up to EventBundleNodeBase
