+++ Core/src/org/sleuthkit/autopsy/timeline/ui/detailview/EventStripeNode.java
@@ -19,17 +19,17 @@
  */
 package org.sleuthkit.autopsy.timeline.ui.detailview;
 
+import java.util.Arrays;
+import java.util.Collection;
 import javafx.event.EventHandler;
 import javafx.geometry.Pos;
 import javafx.scene.control.Button;
-import javafx.scene.control.ContextMenu;
-import javafx.scene.control.MenuItem;
 import javafx.scene.control.OverrunStyle;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseButton;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.VBox;
 import org.apache.commons.lang3.StringUtils;
+import org.controlsfx.control.action.Action;
 import org.controlsfx.control.action.ActionUtils;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.timeline.datamodel.EventCluster;
@@ -83,7 +83,6 @@ final public class EventStripeNode extends EventBundleNodeBase<EventStripe, Even
         }
 
         getChildren().addAll(new VBox(infoHBox, subNodePane));
-        setOnMouseClicked(new MouseClickHandler());
     }
 
     @Override
@@ -146,43 +145,14 @@ final public class EventStripeNode extends EventBundleNodeBase<EventStripe, Even
         }
     }
 
-    /**
-     * event handler used for mouse events on {@link EventStripeNode}s
-     */
-    private class MouseClickHandler implements EventHandler<MouseEvent> {
-
-        private ContextMenu contextMenu;
-
-        @Override
-        public void handle(MouseEvent t) {
-
-            if (t.getButton() == MouseButton.PRIMARY) {
-
-                if (t.isShiftDown()) {
-                    if (chart.selectedNodes.contains(EventStripeNode.this) == false) {
-                        chart.selectedNodes.add(EventStripeNode.this);
-                    }
-                } else if (t.isShortcutDown()) {
-                    chart.selectedNodes.removeAll(EventStripeNode.this);
-                } else {
-                    chart.selectedNodes.setAll(EventStripeNode.this);
-                }
-                t.consume();
-            } else if (t.getButton() == MouseButton.SECONDARY) {
-                ContextMenu chartContextMenu = chart.getChartContextMenu(t);
-                if (contextMenu == null) {
-                    contextMenu = new ContextMenu();
-                    contextMenu.setAutoHide(true);
-
-                    EventDetailsChart.HideDescriptionAction hideClusterAction = chart.new HideDescriptionAction(getDescription(), eventBundle.getDescriptionLoD());
-                    MenuItem hideDescriptionMenuItem = ActionUtils.createMenuItem(hideClusterAction);
-                    contextMenu.getItems().addAll(hideDescriptionMenuItem);
-                    contextMenu.getItems().addAll(chartContextMenu.getItems());
-                }
-                contextMenu.show(EventStripeNode.this, t.getScreenX(), t.getScreenY());
-                t.consume();
-            }
-        }
+    @Override
+    EventHandler<MouseEvent> getDoubleClickHandler() {
+        return mouseEvent -> {
+        };
     }
 
+    @Override
+    Collection<? extends Action> getActions() {
+        return Arrays.asList(chart.new HideDescriptionAction(getDescription(), eventBundle.getDescriptionLoD()));
+    }
 }
